backend/paaster/pyproject.toml
[tool.poetry]
name = "paaster"
version = "2.0.0"
description = ""
authors = ["WardPearce <wardpearce@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
starlite = {extras = ["full"], version = "^1.51.0"}
motor = "^3.1.1"
python-dotenv = "^0.21.1"
bcrypt = "^4.0.1"


[tool.poetry.group.test.dependencies]
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
